requirements.txt
streamlit
groq
python-dotenv
pillow
easyocr
torch
opencv-python-headless
